requirements.txt
Flask
Flask-Cors
PyYAML
pyodbc
pytz
waitress
